5.1. Move utility function to seperate files common and render accordinly
