--- src/utils.js
@@ -1,88 +1,0 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-};
-
-
-/**
- * Returns HTMLElement from template string
- * @param {String} template HTML markup
- * @return {HTMLElement}
- */
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstChild;
-};
-
-
-/**
- * Renders HTMLElement into another HTMLElement
- * @param {HTMLElement} container
- * @param {HTMLElement} element
- * @param {String} place
- */
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
-
-/**
- * Returns a random integer in a range
- * @param {Number} min minimal number (included)
- * @param {Number} max maximal number (included)
- * @return {Number} random number
- */
-const getRandomInteger = (min, max) => {
-  return Math.floor(Math.random() * (max - min + 1) + min);
-};
-
-
-/**
- * Returns random array element
- * @param {Array} array array
- * @return {*} random array element
- */
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomInteger(0, array.length - 1);
-
-  return array[randomIndex];
-};
-
-
-/**
- * Returns value in 2 digit zero padding format
- * @param {Number} value number to pad
- * @return {String} Zero padded value
- */
-const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-
-/**
- * Returns formated date
- * @param {Date} date date to format
- * @return {String} formated string (HH:MM am|pm)
- */
-const formatTime = (date) => {
-  const hours = castTimeFormat(date.getHours() % 12);
-  const minutes = castTimeFormat(date.getMinutes());
-
-  const interval = date.getHours() > 11 ? `PM` : `AM`;
-
-  return `${hours}:${minutes} ${interval}`;
-};
-
-
-export {getRandomInteger};
-export {getRandomArrayItem};
-export {formatTime};
