+++ src/utils/common.js
@@ -1,0 +1,46 @@
+/**
+ * Returns a random integer in a range
+ * @param {Number} min minimal number (included)
+ * @param {Number} max maximal number (included)
+ * @return {Number} random number
+ */
+export const getRandomInteger = (min, max) => {
+  return Math.floor(Math.random() * (max - min + 1) + min);
+};
+
+
+/**
+ * Returns random array element
+ * @param {Array} array array
+ * @return {*} random array element
+ */
+export const getRandomArrayItem = (array) => {
+  const randomIndex = getRandomInteger(0, array.length - 1);
+
+  return array[randomIndex];
+};
+
+
+/**
+ * Returns value in 2 digit zero padding format
+ * @param {Number} value number to pad
+ * @return {String} Zero padded value
+ */
+export const castTimeFormat = (value) => {
+  return value < 10 ? `0${value}` : String(value);
+};
+
+
+/**
+ * Returns formated date
+ * @param {Date} date date to format
+ * @return {String} formated string (HH:MM am|pm)
+ */
+export const formatTime = (date) => {
+  const hours = castTimeFormat(date.getHours() % 12);
+  const minutes = castTimeFormat(date.getMinutes());
+
+  const interval = date.getHours() > 11 ? `PM` : `AM`;
+
+  return `${hours}:${minutes} ${interval}`;
+};
